requirements.txt
torch>=1.1.0
torchvision>=0.3.0
tqdm
dominate
numpy>=1.15.4
opencv-python
Pillow>=5.3.0
PyYAML>=5.1.2
visdom
